Move tests to subdirectories/sub classes.
