+++ NWN.Anvil.Tests/src/main/API/Nui/Bindings/NuiValueTests.cs
@@ -5,22 +5,8 @@ using NUnit.Framework;
 namespace Anvil.Tests.API
 {
   [TestFixture(Category = "API.Nui")]
-  public sealed class NuiBindingTests
+  public class NuiValueTests
   {
-    [Test(Description = "Serializing a NuiBind<string> creates a valid JSON structure.")]
-    public void SerializeNuiBindStringReturnsValidJsonStructure()
-    {
-      NuiBind<string> test = new NuiBind<string>("test");
-      Assert.That(JsonUtility.ToJson(test), Is.EqualTo(@"{""bind"":""test""}"));
-    }
-
-    [Test(Description = "Serializing a NuiBind<NuiRect> creates a valid JSON structure.")]
-    public void SerializeNuiBindNuiRectReturnsValidJsonStructure()
-    {
-      NuiBind<NuiRect> test = new NuiBind<NuiRect>("test");
-      Assert.That(JsonUtility.ToJson(test), Is.EqualTo(@"{""bind"":""test""}"));
-    }
-
     [Test(Description = "Serializing a NuiValue<string> creates a valid JSON structure.")]
     [TestCase("test", @"""test""")]
     [TestCase(null, @"null")]
@@ -103,20 +89,6 @@ namespace Anvil.Tests.API
     {
       NuiValue<List<int>> test = new NuiValue<List<int>>(new List<int> { 1, 2, 3 });
       Assert.That(JsonUtility.ToJson(test), Is.EqualTo(@"[1,2,3]"));
-    }
-
-    [Test(Description = "Deerializing a NuiBind<string> creates a valid JSON structure.")]
-    public void DeserializeNuiBindStringReturnsValidJsonStructure()
-    {
-      NuiBind<string> test = JsonUtility.FromJson<NuiBind<string>>(@"{""bind"":""test""}");
-      Assert.That(test.Key, Is.EqualTo("test"));
-    }
-
-    [Test(Description = "Deerializing a NuiBind<NuiRect> creates a valid JSON structure.")]
-    public void DeserializeNuiBindNuiRectReturnsValidJsonStructure()
-    {
-      NuiBind<NuiRect> test = JsonUtility.FromJson<NuiBind<NuiRect>>(@"{""bind"":""test""}");
-      Assert.That(test.Key, Is.EqualTo("test"));
     }
 
     [Test(Description = "Deerializing a NuiValue<string> creates a valid JSON structure.")]
